requirements.txt
requests
python-dotenv
argparse
pymongo
pymongo[srv]